requirements.txt
asgiref==3.4.1
beautifulsoup4==4.10.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
cloudinary==1.28.0
cryptography==36.0.0
defusedxml==0.7.1
dj-database-url==0.5.0
dj3-cloudinary-storage==0.0.6
Django==3.2.9
django-allauth==0.47.0
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap4==21.1
django-crispy-forms==1.13.0
django-phone-field==1.8.1
django-phonenumber-field==6.0.0
gunicorn==20.1.0
idna==3.3
oauthlib==3.1.1
phonenumberslite==8.12.39
psycopg2==2.9.2
pycparser==2.21
PyJWT==2.3.0
python3-openid==3.2.0
pytz==2021.3
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
soupsieve==2.3.1
sqlparse==0.4.2
urllib3==1.26.7
